+++ routes/web.php
@@ -1,7 +1,8 @@
 <?php
 
-use App\Http\Controllers\ListingController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
+use App\Http\Controllers\ListingController;
 
 /*
 |--------------------------------------------------------------------------
@@ -14,10 +15,16 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [ListingController::class, 'index']);
-Route::get('/jobs/create', [ListingController::class, 'create']);
-Route::post('/jobs', [ListingController::class, 'save']);
-Route::get('/jobs/{job}/edit', [ListingController::class, 'edit']);
-Route::put('/jobs/{job}', [ListingController::class, 'update']);
-Route::delete('/jobs/{job}', [ListingController::class, 'destroy']);
+Route::get('/', [ListingController::class, 'index'])->name('home');
+Route::get('/jobs/create', [ListingController::class, 'create'])->middleware('auth');
+Route::post('/jobs', [ListingController::class, 'save'])->middleware('auth');
+Route::get('/jobs/manage', [ListingController::class, 'manage'])->middleware('auth');
+Route::get('/jobs/{job}/edit', [ListingController::class, 'edit'])->middleware('auth');
+Route::put('/jobs/{job}', [ListingController::class, 'update'])->middleware('auth');
+Route::delete('/jobs/{job}', [ListingController::class, 'destroy'])->middleware('auth');
 Route::get('/jobs/{job}', [ListingController::class, 'show']);
+Route::get('/register', [UserController::class, 'register'])->middleware('guest');
+Route::post('/users', [UserController::class, 'save'])->middleware('guest');
+Route::get('/login', [UserController::class, 'login'])->middleware('guest')->name('login');
+Route::post('/auth', [UserController::class, 'login'])->middleware('guest');
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
